migrated test code into the profile tests
