+++ __tests__/profile.test.js
@@ -2,6 +2,10 @@ import request from "supertest";
 import express from "express";
 import { sequelize } from "../models/user.js";
 import ProfileRoutes from "../routes/profile.js";
+import {
+  getAccountProfilePicture,
+  setAccountProfilePicture,
+} from "../routes/shared-data.js";
 
 const app = express();
 app.set("view engine", "pug");
@@ -105,5 +109,48 @@ describe("POST /profile/change-password", () => {
       });
 
     expect(response.status).toBe(200);
+  });
+});
+
+describe("POST /profile and test profile picture functionality", () => {
+  it("should return 200 if the profile picture can be changed to the next image", async () => {
+    setAccountProfilePicture(0);
+
+    const currentProfilePicture = getAccountProfilePicture();
+    const response = await request(app)
+      .post("/profile")
+      .type("form")
+      .send({ action: "nextPicture" });
+    const response2 = await request(app)
+      .post("/profile")
+      .type("form")
+      .send({ action: "savePicture" });
+
+    expect(response.status).toBe(200);
+    expect(response2.status).toBe(200);
+
+    let updatedProfilePicture = getAccountProfilePicture();
+    expect(updatedProfilePicture).toBe(currentProfilePicture + 1);
+  });
+});
+
+describe("POST /profile and test profile picture functionality", () => {
+  it("should return 200 if the profile picture can be changed to the previous image", async () => {
+    const currentProfilePicture = getAccountProfilePicture();
+
+    const response = await request(app)
+      .post("/profile")
+      .type("form")
+      .send({ action: "prevPicture" });
+    const response2 = await request(app)
+      .post("/profile")
+      .type("form")
+      .send({ action: "savePicture" });
+
+    expect(response.status).toBe(200);
+    expect(response2.status).toBe(200);
+
+    let updatedProfilePicture = getAccountProfilePicture();
+    expect(updatedProfilePicture).toBe(currentProfilePicture - 1);
   });
 });
